saving + loading functionality started

+++ code/VR Art Gallery/Assets/Scripts/Cloud/Models/ArtworkData.cs
@@ -10,22 +10,21 @@ using Postgrest.Models;
 public class ArtworkData : BaseModel
 {
     [PrimaryKey("id")]
-    public int id { get; set; }
-    
+    public long id { get; set; }
     public string title { get; set; }
-    public int owner_id { get; set; }
-  
+    public long owner_id { get; set; }
+
     // Image storage
     public string image_url { get; set; }        // Full resolution artwork
     public string thumbnail_url { get; set; }    // Small preview (512x512)
-    
+
     // Metadata
     public long filesize_bytes { get; set; }
-    
+
     // Timestamps
     public DateTime created_at { get; set; }
     public DateTime updated_at { get; set; }
-    
+
     public ArtworkData()
     {
         created_at = DateTime.UtcNow;
