+++ code/VR Art Gallery/Assets/Scripts/Testing/EditMode/Cloud/Artwork/IArtWorkRepository.cs
@@ -3,5 +3,5 @@ using System.Threading.Tasks;
 public interface IArtworkRepository
 {
     Task<ArtworkData> CreateArtworkAsync(ArtworkData artwork);
-    Task<ArtworkData> GetArtworkAsync(int id);
+    Task<ArtworkData> GetArtworkAsync(long id);
 }
