+++ code/VR Art Gallery/Assets/Scripts/Testing/EditMode/Cloud/Artwork/MockArtworkTests.cs
@@ -3,16 +3,16 @@ using System.Threading.Tasks;
 
 public class MockArtworkRepository : IArtworkRepository
 {
-    private Dictionary<int, ArtworkData> artworks = new Dictionary<int, ArtworkData>();
-    
+    private Dictionary<long, ArtworkData> artworks = new Dictionary<long, ArtworkData>();
+
     public Task<ArtworkData> CreateArtworkAsync(ArtworkData artwork)
     {
-        artwork.id = artworks.Count;
+        artwork.id = (long)artworks.Count;
         artworks[artwork.id] = artwork;
         return Task.FromResult(artwork);
     }
-    
-    public Task<ArtworkData> GetArtworkAsync(int id)
+
+    public Task<ArtworkData> GetArtworkAsync(long id)
     {
         if (artworks.TryGetValue(id, out var artwork))
         {
